Split out the generator and template into new modules
I decided to rethink the generators, and it seemed better to make them
into their own distinct package separate from the RelationalData.
Now a generator gets added to a Mock::Data instance, and Set is a
subclass of generator, and regular subs can be blessed as Gneertor
objects, and the utility method 'inflate_template' is a better
name for a function that only sometimes returns generator objects.