Finished implementation of charset_invlist and parse_charset
Yak-shaving at its finest, I now have code that can parse a Perl
regex charset notation and generate inversion lists from the pieces.
This enables the code to convert a full-unicode charset into an
inversion list much faster than the brute force method, making it
practical to pull random characters out of [:alpha:] and so on in
very modest amounts of processing time and memory usage.

In the event that parsing fails or features are missing on old Perls,
this falls back to the brute force method.