Revise main API of Mock::Data, call/wrap/combine_generators/load_plugin
generators is now a hashref Generator objects, rather than raw
specification.  It's easier to work with when there are fewer datatype
possibilities.  coerce_generator now happens immediately on construction
and assignment.

call is renamed from call_generator, since it might get used a lot and
the long name is awkward.

wrap is a new method that compiles the generator with optional args and
then wraps that in a coderef for fast repeat execution.

combine_generators (formerly merge_generators) is more object-oriented,
and the special case for merging sets is handled by the Set subclass.

GeneratorSub (renamed from SubWrapper) now inherits from Generator
and the constructor is public.