Don't assume compiled generators take params, and don't coerce Set->items
The first big change was that I decided not to bless the compiled
generators, because that implied it is generator which could be further
compiled with new parameters, and that places a large API burden on the
compiled subs.

The next big change was to not automatically assume that items given to
Mock::Data::Set should be coerced into generators.  This allows Set to
contain and return structured data.  Instead, the coercion is performed
deeply by coerce_generator before constructing the Set objects, so the
API remains just as convenient as before.