Rework parse/compilation of Templates
This should be a little more efficient, and adds more capability
to diagnose or extend the syntax.

Also, I decided that named parameters should not be arbitrarily
passed through to the inner templates.  If they were, it would be
impossible to control which passed through.  The template is now
essentially finalized the moment it is specified.