More efficient implementation of _charset_invlist_brute_force
The new design uses perl's /g to iterate a string of characters
instead of iterating characters and invoking the regex engine on
each of them.  This approach uses a chunk of memory, but is more
than twice as fast.  The memory is allocated lazily and should
only affect older perl versions, which were the ones worst hit
by the speed anyway.  Also, fixed a bug where the charset invlist
wasn't getting cached, which made things a lot faster.

This also solves the problem where older perls couldn't compare
non-character codepoints in a regex.